Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -5,7 +5,7 @@ char* welcome() {
 }
 
 int main() {
-    double r;
+    float r;
     float pi=3.14,area;
     scanf("%f",&r);
     area=pi*r*r;
